requirements.txt
appdirs==1.4.4
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
distlib==0.3.1
docutils==0.16
filelock==3.0.12
idna==2.10
importlib-metadata==2.0.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Kivy==1.11.1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.sdl2==0.1.23
Kivy-Garden==0.1.4
numpy==1.19.4
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
Pygments==2.7.2
pypiwin32==223
pywin32==300
requests==2.25.0
six==1.15.0
traitlets==5.0.5
urllib3==1.26.2
virtualenv==20.1.0
wcwidth==0.2.5
zipp==3.4.0
